encrypt/decrypt long messages support
