+++ internal/crypto/encode.go
@@ -1,0 +1,45 @@
+package crypto
+
+import (
+	"crypto/rand"
+	"crypto/rsa"
+	"crypto/sha256"
+	"hash"
+	"io"
+)
+
+// encryptOAEP is a rsa.EncryptOAEP wrapper
+// which encrypts a long messages by splitting them to chunks.
+func encryptOAEP(hash hash.Hash, random io.Reader, public *rsa.PublicKey, msg []byte, label []byte) ([]byte, error) {
+	msgLen := len(msg)
+	step := public.Size() - 2*hash.Size() - 2
+	var encryptedBytes []byte
+
+	for start := 0; start < msgLen; start += step {
+		finish := start + step
+		if finish > msgLen {
+			finish = msgLen
+		}
+
+		encryptedBlockBytes, err := rsa.EncryptOAEP(hash, random, public, msg[start:finish], label)
+		if err != nil {
+			return nil, err
+		}
+
+		encryptedBytes = append(encryptedBytes, encryptedBlockBytes...)
+	}
+
+	return encryptedBytes, nil
+}
+
+// EncryptWithPublicKey encrypts data with public key.
+func EncryptWithPublicKey(msg []byte, pub *rsa.PublicKey) ([]byte, error) {
+	hash := sha256.New()
+
+	ciphertext, err := encryptOAEP(hash, rand.Reader, pub, msg, nil)
+	if err != nil {
+		return nil, err
+	}
+
+	return ciphertext, nil
+}
